# promql: faster range-query of label_replace and label_join