@juliusv 2025-07-09
This seems to have broken the data + query case where series are not duplicate at individual steps of a range query, but become duplicates when combining the outputs of multiple independent range query steps, as reported in https://github.com/prometheus/prometheus/issues/15114